Adicionar tudo ate Crud de Produtos

+++ app/services/usuario_service.ts
@@ -1,39 +1,62 @@
 import Usuario from '#models/usuario'
+import { inject } from '@adonisjs/core'
 import UsuarioPayload from '../payloads/usuarioPayload.js'
 import TelefoneService from './telefone_service.js'
+import UsuarioPutPayload from '../payloads/usuarioPutPayload.js'
+import AlreadyExistsException from '#exceptions/already_exists_exception'
+import isValidCPF from '../utils/isValidCPF.js'
+import ResourceNotFoundException from '#exceptions/resource_not_found_exception'
+import BadRequestException from '#exceptions/bad_request_exception'
 
+@inject()
 export default class UsuarioService {
   constructor(protected telefoneService: TelefoneService) {}
 
-  async encontrarPorCPF(cpf: string) {
-    return await Usuario.findByOrFail('cpf', cpf)
+  async encontrarPorId(id: bigint | undefined) {
+    const usuario = await Usuario.findBy('id', id)
+
+    if (!usuario) {
+      throw new ResourceNotFoundException('Usuário não encontrado')
+    }
+
+    return usuario
   }
 
   async criar(usuario: UsuarioPayload) {
     if (await Usuario.findBy('cpf', usuario.cpf)) {
-      throw new Error('Usuário já cadastrado')
+      throw new AlreadyExistsException('CPF já cadastrado')
+    }
+
+    if (!isValidCPF(usuario.cpf)) {
+      throw new BadRequestException('CPF inválido')
     }
 
     let novoUsuario = new Usuario().merge(usuario)
-    novoUsuario.related('telefones').createMany(usuario.telefones)
+    await novoUsuario.related('telefones').createMany(usuario.telefones)
 
-    return await Usuario.create(usuario)
+    novoUsuario.ativo = true
+    novoUsuario.validado = false
+
+    return await Usuario.create(novoUsuario)
   }
 
-  async atualizar(cpf: string, carga: Partial<Usuario>) {
-    if (!(await Usuario.findBy('cpf', cpf))) {
-      throw new Error('Usuário não encontrado')
+  async atualizar(id: bigint, carga: UsuarioPutPayload) {
+    const usuario = await Usuario.findBy('id', id)
+
+    if (!usuario) {
+      throw new ResourceNotFoundException('Usuário não encontrado')
     }
 
-    const usuario = await Usuario.findByOrFail('cpf', cpf)
-
     usuario.merge(carga)
-
-    await usuario.save()
+    return await usuario.save()
   }
 
-  async deletar(cpf: string) {
-    const usuario = await Usuario.findBy('cpf', cpf)
+  async deletar(id: bigint) {
+    const usuario = await Usuario.findBy('id', id)
+
+    if (!usuario) {
+      throw new ResourceNotFoundException('Usuário não encontrado')
+    }
 
     return await usuario?.delete()
   }
